requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pytesseract==0.3.6
Werkzeug==1.0.1
gunicorn
pytz
